# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 2.0.0
  description: PharmaPOS Pakistan - Complete Pharmacy Management System API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /customers:
    get:
      operationId: getCustomers
      tags: [customers]
      parameters:
        - name: search
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Customers
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Customer"
    post:
      operationId: createCustomer
      tags: [customers]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCustomerInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Customer"
  /customers/{id}:
    get:
      operationId: getCustomer
      tags: [customers]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Customer
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Customer"
    put:
      operationId: updateCustomer
      tags: [customers]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCustomerInput"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Customer"
    delete:
      operationId: deleteCustomer
      tags: [customers]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
...
  /credit:
    get:
      operationId: getCreditSales
      tags: [credit]
      summary: Get all unpaid/partial credit sales (Khata)
      parameters:
        - name: customerId
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Credit sales
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Sale"
  /credit/{saleId}/pay:
    post:
      operationId: recordCreditPayment
      tags: [credit]
      summary: Record a payment against a credit sale
      parameters:
        - name: saleId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreditPaymentInput"
      responses:
        "200":
          description: Payment recorded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Sale"
  /credit/customer/{customerId}:
    get:
      operationId: getCustomerCredit
      tags: [credit]
      summary: Get total outstanding credit for a customer
      parameters:
        - name: customerId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Customer credit summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CustomerCreditSummary"
...
components:
  schemas:
...
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
      required: [success, message]
...
    Customer:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        phone:
          type: string
        email:
          type: string
        address:
          type: string
        cnic:
          type: string
        dateOfBirth:
          type: string
        totalPurchases:
          type: number
        totalCredit:
          type: number
        visitCount:
          type: integer
        createdAt:
          type: string
      required: [id, name, phone, createdAt]
    CreateCustomerInput:
      type: object
      properties:
        name:
          type: string
        phone:
          type: string
        email:
          type: string
        address:
          type: string
        cnic:
          type: string
        dateOfBirth:
          type: string
      required: [name, phone]
    SaleItem:
      type: object
      properties:
        id:
          type: integer
        medicineId:
          type: integer
        medicineName:
          type: string
        medicineUnit:
          type: string
        quantity:
          type: integer
        unitPrice:
          type: number
        discount:
          type: number
        total:
          type: number
      required: [id, medicineId, medicineName, medicineUnit, quantity, unitPrice, discount, total]
    Sale:
      type: object
      properties:
        id:
          type: integer
        invoiceNumber:
          type: string
        customerId:
          type: integer
        customerName:
          type: string
        items:
          type: array
          items:
            $ref: "#/components/schemas/SaleItem"
        subtotal:
          type: number
        discount:
          type: number
        tax:
          type: number
        totalAmount:
          type: number
        paidAmount:
          type: number
        changeAmount:
          type: number
        creditAmount:
          type: number
        paymentStatus:
          type: string
          enum: [paid, credit, partial]
        paymentMethod:
          type: string
          enum: [cash, card, jazzcash, easypaisa, credit]
        status:
          type: string
        notes:
          type: string
        createdAt:
          type: string
      required: [id, invoiceNumber, items, subtotal, discount, tax, totalAmount, paidAmount, changeAmount, creditAmount, paymentStatus, paymentMethod, status, createdAt]
...
    CreditPaymentInput:
      type: object
      properties:
        amountPaid:
          type: number
        paymentMethod:
          type: string
        notes:
          type: string
      required: [amountPaid]
    CustomerCreditSummary:
      type: object
      properties:
        customerId:
          type: integer
        customerName:
          type: string
        totalCredit:
          type: number
        totalPaid:
          type: number
        outstanding:
          type: number
        sales:
          type: array
          items:
            $ref: "#/components/schemas/Sale"
      required: [customerId, customerName, totalCredit, totalPaid, outstanding, sales]
...